Refactor image deletion logic and update IImageService contract
Added a check to prevent image deletion if it's the only one and requires a new image to be added first. Updated `IImageService` interface to include the new `DeleteAsync` method, enhancing service abstraction. Updated `ProductService` to use `IImageService` for dependency consistency.

+++ src/DevExpert.Marketplace.Core/Domain/Images/ImageService.cs
@@ -30,12 +30,20 @@ public class ImageService(
             return;
         }
         
+        var images = await repository.GetImagesByProductIdAsync(image.ProductId.GetValueOrDefault());
+
+        if (images.Count == 1)
+        {
+            notifier.AddNotification(new Notification("Primeiro, insira uma nova imagem."));
+            return;
+        }
+        
         await repository.DeleteAsync(image);
         await repository.SaveChangesAsync();
         
         DeleteImage(image.ProductId.GetValueOrDefault(), image.Name);;
         
-        var images = await repository.GetImagesByProductIdAsync(image.ProductId.GetValueOrDefault());
+        images.Remove(image);
         await ReorderProductImagesDisplayPositionAsync(images);
     }
